+++ packages/server/modules/core/helpers/graphTypes.ts
@@ -67,6 +67,7 @@ export type ProjectGraphQLReturn = StreamRecord &
     | 'modelChildrenTree'
     | 'viewerResources'
     | 'visibility'
+    | 'allowPublicComments'
   > & {
     /**
      * Some queries resolve the role, some don't. If role isn't returned, no worries, it'll
